refactoring sezione Employee

+++ src/main/java/co/develhope/progettoDiGruppoRecap/Repository/EmployeeRepository.java
@@ -4,6 +4,6 @@ import co.develhope.progettoDiGruppoRecap.Entity.EmployeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface EmployeeRepository extends JpaRepository <EmployeeEntity , Integer> {
+public interface EmployeeRepository extends JpaRepository <EmployeeEntity , Long> {
     //Ricerca dipendenti per stipendio
 }
